[StoreItem] downloadImage() 추가
